Updated JavaScript for Matches html <p>
This commit changes the query’s below the jumbotron to which all 3 of
the query input/form text fields are responsive. You can enter
different query searches such as Unbroken into one of them and War and
peace in the other and they’ll both show differing numbers for
indication of the number of different results there is.