Updated Navigational Bar Links for SignUp/Login
Created a registration script, previously used it for a website I
worked on in the past. The registration.php successfully creates users
to be inserted into the database, and makes simple checks for proper
input formatting for things such as email. Will update login check
later this weekend/week, but for now it will be useful to use it to
fiddle around with the database.